requirements.txt
requests
lxml
python-telegram-bot
pyyaml